리뷰 등록

+++ src/main/java/com/jpa/exercise/repository/HospitalRepository.java
@@ -4,5 +4,5 @@ package com.jpa.exercise.repository;
 import com.jpa.exercise.domain.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital,Long> {
+public interface HospitalRepository extends JpaRepository<Hospital,Integer> {
 }
